producing correct output, still need to cleanup and test visitor functions

+++ Proj03/Employee.h
@@ -102,7 +102,7 @@ public:
         
     }
     
-    unsigned int getSalary() const{
+    unsigned long getSalary() const{
         
         return m_Salary;
     }
